todo-v3/requirements.txt
asgiref
Django
django-crontab
djangorestframework
djangorestframework-simplejwt
numpy
pandas
Pillow
PyJWT
python-dateutil
python-decouple
pytz
six
sqlparse
